refactor: config parser
