--- vocascan.config1.js
@@ -1,5 +1,0 @@
-module.exports = {
-  server: {
-    port: 7643,
-  },
-};
